requirements.txt
alembic==1.15.2
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Deprecated==1.2.18
ecdsa==0.19.1
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==9.3.0
ordered-set==4.1.0
packaging==25.0
pyasn1==0.4.8
Pygments==2.19.1
PyMySQL==1.1.1
python-jose==3.4.0
rich==13.9.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
